Bug fixin' and tables in views
Shows users in all.haml and items in active.haml now in table
like "my active items".

Added check to make sure that the default picture is not removed.
Set description to "" when an item is created.

modified:   trade/app/models/item.rb
modified:   trade/app/views/item/active.haml
modified:   trade/app/views/user/all.haml

+++ trade/app/models/item.rb
@@ -30,6 +30,7 @@ module Models
       item.price = price
       item.active = false
       item.owner = owner
+      item.description = ""
       item.picture = "/images/items/default_item.png"
       item
     end
@@ -84,7 +85,9 @@ module Models
     def clear
       System.instance.remove_item(self.id)
       begin
-        File.delete("#{self.picture.sub("/images", "./public/images")}")
+        unless self.picture =~ /default_item\.png$/
+          File.delete("#{self.picture.sub("/images", "./public/images")}")
+        end
       rescue => e
         puts(e)
         puts("Picture does not exist on #{self.picture.sub("/images", "./public/images")}")
